flatten db
