--- crates/engine/src/db/mod.rs
@@ -1,144 +1,0 @@
-use std::sync::Arc;
-
-use enum_map::EnumMap;
-use spacetimedb_schema::reducer_name::ReducerName;
-
-use crate::metrics::ExecutionCounters;
-use spacetimedb_datastore::execution_context::WorkloadType;
-use spacetimedb_datastore::{locking_tx_datastore::datastore::TxMetrics, traits::TxData};
-
-mod durability;
-pub mod persistence;
-pub mod relational_db;
-pub mod snapshot;
-pub mod update;
-
-/// Whether SpacetimeDB is run in memory, or persists objects and
-/// a message log to disk.
-#[derive(Clone, Copy)]
-pub enum Storage {
-    /// The object store is in memory, and no message log is kept.
-    Memory,
-
-    /// The object store is persisted to disk, and a message log is kept.
-    Disk,
-}
-
-/// Internal database config parameters
-#[derive(Clone, Copy)]
-pub struct Config {
-    /// Specifies the object storage model.
-    pub storage: Storage,
-    /// Specifies the page pool max size in bytes.
-    pub page_pool_max_size: Option<usize>,
-}
-
-/// A message that is processed by the [`spawn_metrics_recorder`] actor.
-/// We use a separate task to record metrics to avoid blocking transactions.
-pub struct MetricsMessage {
-    /// The reducer the produced these metrics.
-    reducer: Option<ReducerName>,
-    /// Metrics from a mutable transaction.
-    metrics_for_writer: Option<TxMetrics>,
-    /// Metrics from a read-only transaction.
-    /// A message may have metrics for both types of transactions,
-    /// because metrics for a reducer and its subscription updates are recorded together.
-    metrics_for_reader: Option<TxMetrics>,
-    /// The row updates for an immutable transaction.
-    /// Needed for insert and delete counters.
-    tx_data: Option<Arc<TxData>>,
-    /// Cached metrics counters for each workload type.
-    counters: Arc<EnumMap<WorkloadType, ExecutionCounters>>,
-}
-
-/// The handle used to send work to the tx metrics recorder.
-#[derive(Clone)]
-pub struct MetricsRecorderQueue {
-    tx: spacetimedb_runtime::sync::mpsc::UnboundedSender<MetricsMessage>,
-}
-
-impl MetricsRecorderQueue {
-    pub fn send_metrics(
-        &self,
-        reducer: Option<ReducerName>,
-        metrics_for_writer: Option<TxMetrics>,
-        metrics_for_reader: Option<TxMetrics>,
-        tx_data: Option<Arc<TxData>>,
-        counters: Arc<EnumMap<WorkloadType, ExecutionCounters>>,
-    ) {
-        if let Err(err) = self.tx.send(MetricsMessage {
-            reducer,
-            metrics_for_writer,
-            metrics_for_reader,
-            tx_data,
-            counters,
-        }) {
-            log::warn!("failed to send metrics: {err}");
-        }
-    }
-}
-
-fn record_metrics(
-    MetricsMessage {
-        reducer,
-        metrics_for_writer,
-        metrics_for_reader,
-        tx_data,
-        counters,
-    }: MetricsMessage,
-) {
-    if let Some(tx_metrics) = metrics_for_writer {
-        tx_metrics.report(
-            // If row updates are present,
-            // they will always belong to the writer transaction.
-            tx_data.as_deref(),
-            reducer.as_ref(),
-            |wl| &counters[wl],
-        );
-    }
-    if let Some(tx_metrics) = metrics_for_reader {
-        tx_metrics.report(
-            // If row updates are present,
-            // they will never belong to the reader transaction.
-            // Passing row updates here will most likely panic.
-            None,
-            reducer.as_ref(),
-            |wl| &counters[wl],
-        );
-    }
-}
-
-/// The metrics recorder is a side channel that the main database thread forwards metrics to.
-/// While we want to avoid unnecessary compute on the critical path, communicating with other
-/// threads is not free, and for this case in particular waking a parked task is not free.
-///
-/// Previously, each tx would send its metrics to the recorder task. As soon as the recorder
-/// task `recv`d a message, it would update the counters and gauges, and immediately wait for
-/// the next tx's message. This meant that the tx would need to be more expensive than the
-/// recording of its metrics in order for the recorder task not to be parked on `recv` when
-/// the tx would `send` its metrics. This would obviously never be the case, and so each `send`
-/// would incur the overhead of waking the task.
-///
-/// To mitigate this, we now record metrics, for potentially many transactions, periodically
-/// every 5ms.
-const TX_METRICS_RECORDING_INTERVAL: std::time::Duration = std::time::Duration::from_millis(5);
-
-/// Spawns a task for recording transaction metrics.
-/// Returns the handle for pushing metrics to the recorder.
-pub fn spawn_tx_metrics_recorder(
-    handle: &spacetimedb_runtime::Handle,
-) -> (MetricsRecorderQueue, spacetimedb_runtime::AbortHandle) {
-    let handle_clone = handle.clone();
-    let (tx, mut rx) = spacetimedb_runtime::sync::mpsc::unbounded_channel();
-    let abort_handle = handle
-        .spawn(async move {
-            loop {
-                handle_clone.sleep(TX_METRICS_RECORDING_INTERVAL).await;
-                while let Ok(metrics) = rx.try_recv() {
-                    record_metrics(metrics);
-                }
-            }
-        })
-        .abort_handle();
-    (MetricsRecorderQueue { tx }, abort_handle)
-}
