+++ crates/engine/src/relational_db.rs
@@ -1,8 +1,111 @@
-use crate::db::durability::{request_durability, spawn_close as spawn_durability_close};
-use crate::db::MetricsRecorderQueue;
+use crate::durability::{request_durability, spawn_close as spawn_durability_close};
 use crate::error::{DBError, RestoreSnapshotError};
 use crate::metrics::ExecutionCounters;
 use crate::metrics::ENGINE_METRICS;
+
+/// Whether SpacetimeDB is run in memory, or persists objects and
+/// a message log to disk.
+#[derive(Clone, Copy)]
+pub enum Storage {
+    /// The object store is in memory, and no message log is kept.
+    Memory,
+
+    /// The object store is persisted to disk, and a message log is kept.
+    Disk,
+}
+
+/// Internal database config parameters
+#[derive(Clone, Copy)]
+pub struct Config {
+    /// Specifies the object storage model.
+    pub storage: Storage,
+    /// Specifies the page pool max size in bytes.
+    pub page_pool_max_size: Option<usize>,
+}
+
+/// A message that is processed by the [`spawn_metrics_recorder`] actor.
+/// We use a separate task to record metrics to avoid blocking transactions.
+pub struct MetricsMessage {
+    /// The reducer the produced these metrics.
+    reducer: Option<ReducerName>,
+    /// Metrics from a mutable transaction.
+    metrics_for_writer: Option<TxMetrics>,
+    /// Metrics from a read-only transaction.
+    /// A message may have metrics for both types of transactions,
+    /// because metrics for a reducer and its subscription updates are recorded together.
+    metrics_for_reader: Option<TxMetrics>,
+    /// The row updates for an immutable transaction.
+    /// Needed for insert and delete counters.
+    tx_data: Option<Arc<TxData>>,
+    /// Cached metrics counters for each workload type.
+    counters: Arc<EnumMap<WorkloadType, ExecutionCounters>>,
+}
+
+/// The handle used to send work to the tx metrics recorder.
+#[derive(Clone)]
+pub struct MetricsRecorderQueue {
+    tx: spacetimedb_runtime::sync::mpsc::UnboundedSender<MetricsMessage>,
+}
+
+impl MetricsRecorderQueue {
+    pub fn send_metrics(
+        &self,
+        reducer: Option<ReducerName>,
+        metrics_for_writer: Option<TxMetrics>,
+        metrics_for_reader: Option<TxMetrics>,
+        tx_data: Option<Arc<TxData>>,
+        counters: Arc<EnumMap<WorkloadType, ExecutionCounters>>,
+    ) {
+        if let Err(err) = self.tx.send(MetricsMessage {
+            reducer,
+            metrics_for_writer,
+            metrics_for_reader,
+            tx_data,
+            counters,
+        }) {
+            log::warn!("failed to send metrics: {err}");
+        }
+    }
+}
+
+fn record_metrics(
+    MetricsMessage {
+        reducer,
+        metrics_for_writer,
+        metrics_for_reader,
+        tx_data,
+        counters,
+    }: MetricsMessage,
+) {
+    if let Some(tx_metrics) = metrics_for_writer {
+        tx_metrics.report(tx_data.as_deref(), reducer.as_ref(), |wl| &counters[wl]);
+    }
+    if let Some(tx_metrics) = metrics_for_reader {
+        tx_metrics.report(None, reducer.as_ref(), |wl| &counters[wl]);
+    }
+}
+
+const TX_METRICS_RECORDING_INTERVAL: std::time::Duration = std::time::Duration::from_millis(5);
+
+/// Spawns a task for recording transaction metrics.
+/// Returns the handle for pushing metrics to the recorder.
+pub fn spawn_tx_metrics_recorder(
+    handle: &spacetimedb_runtime::Handle,
+) -> (MetricsRecorderQueue, spacetimedb_runtime::AbortHandle) {
+    let handle_clone = handle.clone();
+    let (tx, mut rx) = spacetimedb_runtime::sync::mpsc::unbounded_channel();
+    let abort_handle = handle
+        .spawn(async move {
+            loop {
+                handle_clone.sleep(TX_METRICS_RECORDING_INTERVAL).await;
+                while let Ok(metrics) = rx.try_recv() {
+                    record_metrics(metrics);
+                }
+            }
+        })
+        .abort_handle();
+    (MetricsRecorderQueue { tx }, abort_handle)
+}
 use crate::util::asyncify;
 use anyhow::{anyhow, Context};
 use enum_map::EnumMap;
@@ -1830,8 +1933,8 @@ fn default_row_count_fn(db: Identity) -> RowCountFn {
 
 #[cfg(any(test, feature = "test"))]
 pub mod tests_utils {
-    use crate::db::snapshot;
-    use crate::db::snapshot::SnapshotWorker;
+    use crate::snapshot;
+    use crate::snapshot::SnapshotWorker;
 
     use super::*;
     use core::ops::Deref;
@@ -2331,7 +2434,7 @@ mod tests {
 
     use super::tests_utils::begin_mut_tx;
     use super::*;
-    use crate::db::relational_db::tests_utils::{begin_tx, create_view_for_test, insert, make_snapshot, TestDB};
+    use crate::relational_db::tests_utils::{begin_tx, create_view_for_test, insert, make_snapshot, TestDB};
     use anyhow::bail;
     use bytes::Bytes;
     use commitlog::payload::txdata;
