+++ crates/engine/src/rls.rs
@@ -1,4 +1,4 @@
-use crate::sql::ast::SchemaViewer;
+use crate::ast::SchemaViewer;
 use spacetimedb_datastore::locking_tx_datastore::state_view::StateView;
 use spacetimedb_datastore::locking_tx_datastore::MutTxId;
 use spacetimedb_expr::check::parse_and_type_sub;
