+++ crates/engine/src/snapshot.rs
