--- crates/engine/src/sql/mod.rs
@@ -1,1 +1,0 @@
-pub mod ast;
